refactor(benchmark-slicer): rename the hosting file to better reflect what is in it
